Changes.. had to make get_a_number return strings because the times in a the target file seem to be larger than the max float? Maybe? Trying double had similar failure. at any rate, since I'm not performing arithmatic on these numbers, it doesn't matter if they are stored as strings/float and at the very least this guarantees they won't be truncated.
Another problem is that the stored time is in microseconds (which prob
causes the above problem) to get any useful time out of them, I would need
to divide by 1 million to get the unixepoch.

In sqlite I would just need to do:
	select datetime(mtime/1000000,'unixepoch','localtime') from ..

To get it.

+++ json_adt/jso.h
@@ -17,23 +17,23 @@ struct Jso {
 	JType t;
 	union {
 		string* s;
-		float f;
+		double f;
 		vector<Jso*>* a;
 		map<string,Jso*>* m;
 	} x;
-	explicit Jso(float v);
+	explicit Jso(double v);
 	explicit Jso(const string& v);
 	explicit Jso(JType j);
 	~Jso();
 	void key_value(const string& k,Jso* v);
 	void key_value(const string& k,const string& v);
-	void key_value(const string& k,const float v);
+	void key_value(const string& k,const double v);
 	void key_value(const string& k,JType vt);
-	void add_value(const float v);
+	void add_value(const double v);
 	void add_value(const string& v);
 	void add_value(const JType vt);
 	void add_value(Jso* v);
-	void set_value(float v);
+	void set_value(double v);
 	void set_value(const string& v);
 	friend ostream& operator<<(ostream& out,const Jso& J);
 	Jso* key_value(const string& k);
